requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
frozenlist==1.3.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
multidict==6.0.2
pyee==9.0.4
slackclient==2.9.3
slackeventsapi==3.0.1
typing-extensions==4.1.1
Werkzeug==2.0.3
yarl==1.7.2
pytest==7.0.1
waitress==2.0.0